Изм - 1.7.2 работа над организацией
